add creates instance test util
