+++ src/decorators/router.ts
@@ -2,19 +2,13 @@ import { container as defaultContainer } from "../container/container";
 import { setRouter } from "../container/utils/setRouter";
 import { validateContainer } from "../container/utils/validateContainer";
 import { setInjectables } from "./inject/setInjectables";
-import { AnnotatedRouter } from "./types";
+import { AnnotatedRouter, Class } from "./types";
 import { getMetadata } from "./utils/getMetadata";
 import { validateKind } from "./utils/validateKind";
 
 export const Router =
-  <
-    Class extends new (...args: unknown[]) => AnnotatedRouter = new (
-      ...args: unknown[]
-    ) => AnnotatedRouter
-  >(
-    container = defaultContainer
-  ) =>
-  (constructor: Class, context: ClassDecoratorContext<Class>) => {
+  <T extends Class<AnnotatedRouter>>(container = defaultContainer) =>
+  (constructor: T, context: ClassDecoratorContext<T>) => {
     validateContainer(container);
 
     const annotationName = `@${Router.name}`;
